PinnedConcurrentCol migration
`ConcurrentBag` now is only a wrapper around a `PinnedConcurrentCol` with a specific `ConcurrentState` implementation.
* `State` defines and implements the unique properties of the concurrent bag.
* `ConcurrentBag` implements `Deref` and `DerefMut` for `PinnedConcurrentCol`.
* Documentation is revised.
* Benchmarks are repeated, performance has not changed (as expected).
