+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "orx-concurrent-bag"
-version = "1.10.0"
+version = "1.11.0"
 edition = "2021"
 authors = ["orxfun <orx.ugur.arikan@gmail.com>"]
 description = "An efficient, convenient and lightweight grow-only concurrent data structure allowing high performance concurrent collection."
@@ -10,9 +10,10 @@ keywords = ["concurrency", "bag", "data-structures", "atomic", "lock-free"]
 categories = ["data-structures", "concurrency", "rust-patterns"]
 
 [dependencies]
-orx-fixed-vec = "2.7"
-orx-pinned-vec = "2.7"
-orx-split-vec = "2.8"
+orx-fixed-vec = "2.8"
+orx-pinned-concurrent-col = "1.0"
+orx-pinned-vec = "2.8"
+orx-split-vec = "2.9"
 
 [dev-dependencies]
 criterion = "0.5.1"
