+++ src/prelude.rs
@@ -1,4 +1,5 @@
 pub use crate::bag::ConcurrentBag;
 pub use orx_fixed_vec::FixedVec;
+pub use orx_pinned_concurrent_col::PinnedConcurrentCol;
 pub use orx_pinned_vec::{CapacityState, PinnedVec, PinnedVecGrowthError};
 pub use orx_split_vec::{Doubling, Linear, Recursive, SplitVec};
